combine fs.ReadStream and fs.WriteStream into node/internal/fs/streams.ts
