+++ node/internal/fs/streams.ts
@@ -1,10 +1,12 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 // Copyright Joyent, Inc. and Node.js contributors. All rights reserved. MIT license.
 
-import { ERR_INVALID_ARG_TYPE } from "../errors.ts";
+import { ERR_INVALID_ARG_TYPE, ERR_OUT_OF_RANGE } from "../errors.ts";
 import { deprecate, kEmptyObject } from "../util.mjs";
 import { validateFunction, validateInteger } from "../validators.mjs";
+import { errorOrDestroy } from "../streams/destroy.mjs";
 import { open as fsOpen, type openFlags } from "../../_fs/_fs_open.ts";
+import { read as fsRead } from "../../_fs/_fs_read.ts";
 // @deno-types="../../_fs/_fs_write.d.ts""
 import { write as fsWrite } from "../../_fs/_fs_write.mjs";
 // @deno-types="../../_fs/_fs_writev.d.ts""
@@ -17,8 +19,8 @@ import {
   getValidatedFd,
   validatePath,
 } from "./utils.mjs";
-import { Writable } from "../../stream.ts";
-import { WritableOptions } from "../../_stream.d.ts";
+import { finished, Readable, Writable } from "../../stream.ts";
+import { ReadableOptions, WritableOptions } from "../../_stream.d.ts";
 import { toPathIfFileURL } from "../url.ts";
 import { BufferEncoding } from "../../_global.d.ts";
 import { nextTick } from "../../_next_tick.ts";
@@ -29,11 +31,13 @@ const kFs = Symbol("kFs");
 
 interface FS {
   open?: typeof fsOpen;
+  read?: typeof fsRead;
   write?: typeof fsWrite;
   writev?: typeof fsWritev;
   close?: typeof fsClose;
 }
-interface WriteStreamOptions {
+
+interface StreamOptions {
   flags?: openFlags;
   encoding?: BufferEncoding;
   fd?: number;
@@ -43,8 +47,14 @@ interface WriteStreamOptions {
   start?: number;
   fs?: FS | null;
 }
+interface ReadStreamOptions extends StreamOptions {
+  end?: number;
+  highWaterMark?: number;
+}
+// deno-lint-ignore no-empty-interface
+interface WriteStreamOptions extends StreamOptions {}
 
-export interface WriteStream extends Writable {
+interface Stream {
   fd: number | null;
   [kFs]: FS;
   path?: string | Buffer;
@@ -52,22 +62,32 @@ export interface WriteStream extends Writable {
   mode?: number;
   start?: number;
   pos?: number;
-  bytesWritten: number;
   [kIsPerformingIO]: boolean;
   autoClose: boolean;
 }
+export interface ReadStream extends Readable, Stream {
+  end: number;
+  bytesRead: number;
+}
+export interface WriteStream extends Writable, Stream {
+  bytesWritten: number;
+}
+type EitherStream = ReadStream | WriteStream;
 
-function _construct(this: WriteStream, callback: (err?: Error | null) => void) {
-  const stream = this as WriteStream & { open: () => void };
+function _construct(
+  this: EitherStream,
+  callback: (err?: Error | null) => void,
+) {
+  const stream = this as EitherStream & { open: () => void };
   if (typeof stream.fd === "number") {
     callback();
     return;
   }
 
-  if (stream.open !== openWriteFs) {
+  if (stream.open !== openWriteFs && stream.open !== openReadFs) {
     // Backwards compat for monkey patching open().
     const orgEmit = stream.emit;
-    stream.emit = function (this: WriteStream, ...args) {
+    stream.emit = function (this: EitherStream, ...args) {
       if (args[0] === "open") {
         this.emit = orgEmit;
         callback();
@@ -100,7 +120,7 @@ function _construct(this: WriteStream, callback: (err?: Error | null) => void) {
 }
 
 function close(
-  stream: WriteStream,
+  stream: EitherStream,
   err: Error | null,
   cb: (err?: Error | null) => void,
 ) {
@@ -115,21 +135,229 @@ function close(
 }
 
 function importFd(
-  stream: WriteStream,
-  options: WriteStreamOptions & WritableOptions,
+  stream: EitherStream,
+  options: StreamOptions,
 ) {
   if (typeof options.fd === "number") {
     // When fd is a raw descriptor, we must keep our fingers crossed
     // that the descriptor won't get closed, or worse, replaced with
     // another one
     // https://github.com/nodejs/node/issues/35862
-    stream[kFs] = options.fs ||
-      { write: fsWrite, writev: fsWritev, close: fsClose };
+    if (stream instanceof ReadStream) {
+      stream[kFs] = options.fs || { read: fsRead, close: fsClose };
+    }
+    if (stream instanceof WriteStream) {
+      stream[kFs] = options.fs ||
+        { write: fsWrite, writev: fsWritev, close: fsClose };
+    }
     return options.fd;
   }
 
   throw new ERR_INVALID_ARG_TYPE("options.fd", ["number"], options.fd);
 }
+
+export function ReadStream(
+  this: ReadStream | unknown,
+  path: string | Buffer | URL,
+  options?: BufferEncoding | (ReadStreamOptions & ReadableOptions),
+): ReadStream {
+  if (!(this instanceof ReadStream)) {
+    // deno-lint-ignore ban-ts-comment
+    // @ts-ignore
+    return new ReadStream(path, options);
+  }
+
+  const self = this as ReadStream;
+
+  // A little bit bigger buffer and water marks by default
+  options = copyObject(getOptions(options, kEmptyObject));
+  if (options.highWaterMark === undefined) {
+    options.highWaterMark = 64 * 1024;
+  }
+
+  if (options.autoDestroy === undefined) {
+    options.autoDestroy = false;
+  }
+
+  if (options.fd == null) {
+    self.fd = null;
+    self[kFs] = options.fs || { open: fsOpen, read: fsRead, close: fsClose };
+    validateFunction(self[kFs].open, "options.fs.open");
+
+    // Path will be ignored when fd is specified, so it can be falsy
+    self.path = toPathIfFileURL(path);
+    self.flags = options.flags === undefined ? "r" : options.flags;
+    self.mode = options.mode === undefined ? 0o666 : options.mode;
+
+    validatePath(self.path);
+  } else {
+    self.fd = getValidatedFd(importFd(self, options));
+  }
+
+  options.autoDestroy = options.autoClose === undefined
+    ? true
+    : options.autoClose;
+
+  validateFunction(self[kFs].read, "options.fs.read");
+
+  if (options.autoDestroy) {
+    validateFunction(self[kFs].close, "options.fs.close");
+  }
+
+  self.start = options.start;
+  self.end = options.end ?? Infinity;
+  self.pos = undefined;
+  self.bytesRead = 0;
+  self[kIsPerformingIO] = false;
+
+  if (self.start !== undefined) {
+    validateInteger(self.start, "start", 0);
+
+    self.pos = self.start;
+  }
+
+  if (self.end !== Infinity) {
+    validateInteger(self.end, "end", 0);
+
+    if (self.start !== undefined && self.start > self.end) {
+      throw new ERR_OUT_OF_RANGE(
+        "start",
+        `<= "end" (here: ${self.end})`,
+        self.start,
+      );
+    }
+  }
+
+  Reflect.apply(Readable, self, [options]);
+
+  return self;
+}
+
+Object.setPrototypeOf(ReadStream.prototype, Readable.prototype);
+Object.setPrototypeOf(ReadStream, Readable);
+
+Object.defineProperty(ReadStream.prototype, "autoClose", {
+  get() {
+    return this._readableState.autoDestroy;
+  },
+  set(val) {
+    this._readableState.autoDestroy = val;
+  },
+});
+
+const openReadFs = deprecate(
+  function () {
+    // Noop.
+  },
+  "ReadStream.prototype.open() is deprecated",
+  "DEP0135",
+);
+ReadStream.prototype.open = openReadFs;
+
+ReadStream.prototype._construct = _construct;
+
+ReadStream.prototype._read = async function (this: ReadStream, n: number) {
+  n = this.pos !== undefined
+    ? Math.min(this.end - this.pos + 1, n)
+    : Math.min(this.end - this.bytesRead + 1, n);
+
+  if (n <= 0) {
+    this.push(null);
+    return;
+  }
+
+  const buf = Buffer.allocUnsafeSlow(n);
+
+  let error: Error | null = null;
+  let bytesRead: number | null = null;
+  let buffer: Buffer | undefined = undefined;
+
+  this[kIsPerformingIO] = true;
+
+  await new Promise((resolve) => {
+    this[kFs]
+      .read!(
+        this.fd!,
+        buf,
+        0,
+        n,
+        this.pos ?? null,
+        (_er, _bytesRead, _buf) => {
+          error = _er;
+          bytesRead = _bytesRead;
+          buffer = _buf;
+          return resolve(true);
+        },
+      );
+  });
+
+  this[kIsPerformingIO] = false;
+
+  // Tell ._destroy() that it's safe to close the fd now.
+  if (this.destroyed) {
+    this.emit(kIoDone, error);
+    return;
+  }
+
+  if (error) {
+    errorOrDestroy(this, error);
+  } else if (
+    typeof bytesRead === "number" &&
+    bytesRead > 0
+  ) {
+    if (this.pos !== undefined) {
+      this.pos += bytesRead;
+    }
+
+    this.bytesRead += bytesRead;
+
+    if (bytesRead !== buffer!.length) {
+      // Slow path. Shrink to fit.
+      // Copy instead of slice so that we don't retain
+      // large backing buffer for small reads.
+      const dst = Buffer.allocUnsafeSlow(bytesRead);
+      buffer!.copy(dst, 0, 0, bytesRead);
+      buffer = dst;
+    }
+
+    this.push(buffer);
+  } else {
+    this.push(null);
+  }
+};
+
+ReadStream.prototype._destroy = function (
+  this: ReadStream,
+  err: Error | null,
+  cb: (err?: Error | null) => void,
+) {
+  // Usually for async IO it is safe to close a file descriptor
+  // even when there are pending operations. However, due to platform
+  // differences file IO is implemented using synchronous operations
+  // running in a thread pool. Therefore, file descriptors are not safe
+  // to close while used in a pending read or write operation. Wait for
+  // any pending IO (kIsPerformingIO) to complete (kIoDone).
+  if (this[kIsPerformingIO]) {
+    this.once(kIoDone, (er) => close(this, err || er, cb));
+  } else {
+    close(this, err, cb);
+  }
+};
+
+ReadStream.prototype.close = function (
+  this: ReadStream,
+  cb: (err?: Error | null) => void,
+) {
+  if (typeof cb === "function") finished(this, cb);
+  this.destroy();
+};
+
+Object.defineProperty(ReadStream.prototype, "pending", {
+  get() {
+    return this.fd === null;
+  },
+  configurable: true,
+});
 
 export function WriteStream(
   this: WriteStream | unknown,
@@ -365,6 +593,15 @@ Object.defineProperty(WriteStream.prototype, "pending", {
   },
   configurable: true,
 });
+
+export function createReadStream(
+  path: string | Buffer | URL,
+  options?: BufferEncoding | ReadStreamOptions,
+): ReadStream {
+  // deno-lint-ignore ban-ts-comment
+  // @ts-ignore
+  return new ReadStream(path, options);
+}
 
 export function createWriteStream(
   path: string | Buffer | URL,
